adds tests for aliased built-in types nullability
